requirements.txt
astroid==2.4.2
beautifulsoup4==4.6.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
idna==2.10
isort==5.7.0
JPype1==1.2.1
kiwisolver @ file:///D:/bld/kiwisolver_1610099953470/work
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.6.3
matplotlib @ file:///D:/bld/matplotlib-suite_1617212922200/work
mccabe==0.6.1
numpy @ file:///D:/bld/numpy_1616933442622/work
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
Pillow @ file:///D:/bld/pillow_1617818313335/work
pylint==2.6.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
toml==0.10.2
tornado @ file:///D:/bld/tornado_1610094910414/work
tweepy==3.10.0
urllib3==1.26.4
wincertstore==0.2
wordcloud @ file:///D:/bld/wordcloud_1605555768540/work
wrapt==1.12.1
